DLT: Fix warning with bitshifting an integer 32-bit
Shen shifting an integer 32-bits, the results are undefined (likely zero), so the type must be promoted to 64-bit first for this to work.

+++ TraceReader.Dlt/DltTraceReader/Dlt/Control/SyncTimeStampResponseDecoder.cs
@@ -27,7 +27,7 @@
             uint secLow = unchecked((uint)BitOperations.To32Shift(buffer[9..13], !msbf));
             uint secHigh = unchecked((ushort)BitOperations.To16Shift(buffer[13..15], !msbf));
 
-            long sec = (secHigh << 32) | secLow;  // Number of seconds since 1/1/1970
+            long sec = ((long)secHigh << 32) | secLow;  // Number of seconds since 1/1/1970
             long nsTicks = ns / 100;         // TicksPerMillisecond = 10000. i.e. 100ns per tick.
                                              // Manual calculation to avoid int overflow.
             DateTimeOffset timeStamp = DateTimeOffset.FromUnixTimeSeconds(sec).AddTicks(nsTicks);
